Squash commit for 3-tier CI testing
## Summary
This MR introduces a 3-tier testing system that changes the way jobs are run in the CI. Jobs are now bucketed into the pre-submit, post-submit and nightly flows/tiers. Depending on how the pipeline was triggered, only one of these flows would be activated and only the jobs marked to be part of that flow would be kicked off.

As part of these new tiers, GC tests have been moved out of the pre-submit tier. The jobs running on Pascal nodes have been moved to the nightly tier, whereas the jobs running on A100/A40 nodes have moved to the post-submit tier.

In addition, the behavior of toggle/control variables is also reversed and entire workflow section has been refactored to support the main change.

## Implementation
- This is done by using the TRIGGER_PIPELINE variable, and manipulating it to activate jobs to be included in each flow.
- The toggle variables are all set to 0 (instead of 1), and they are now used to forcibly include jobs in a flow that wouldn't otherwise be included.

### Note
The implementation may have broken `pipeline_ctl`, so we need to refactor that to make it work with this change in a separate patch. 

## Testing
Running CI pipelines with different values assigned to TRIGGER_PIPELINE and ensuring the right jobs are triggered. The rest of the behavior would need to be tested in production after this MR merges.

## Tracking
https://nvbugspro.nvidia.com/bug/5241107

added pipeline flows

See merge request nccl/nccl!1259