Refactor ncclIbCreateQp to configure every QP with the correct properties
Each QP has different properties depending on its usage. For example, a send
QP does not need a receive queue. Another example is A QP used for GPU flush
operations may need a smaller send queue than a QP used for data transfer.

The TRACE message printed in the ncclIbCreateQp function was moved to be
printed in the caller functions. The printed message used information that was
out of the scope of the ncclIbCreateQp function and to allow calling the
ncclIbCreateQp with the correct and proper input arguments, the print message
was moved to the caller.