graph/search: add PCIe based rail id for MNNVL systems
The rail id is used to indicate to NCCL which network devices are connected through the same rail.
This is used to enforce NCCL_CROSS_NIC=0 in the graph search.

When comparing devices on the same host, we use the guid + port to detect same rails.
The PCIe path is not used because it's too restrictive: a device (such as CX8) could have different PCIe path for a single guid.
This would lead NCCL not to select an exit NIC while it could/should.

For multinode NVLink systems, however, the rail can be seen either as being "per-domain" or "per-host".
User can choose one mode or the other with NCCL_MNNVL_RAIL_PER_HOST=0/1.
The default being 0 for 2.28 in order to preserve the existing behavior.

In a "per-domain" rail configuration, each device in the NVLink domain is assumed to be on a different rail.
NCCL uses the GUID + port ID to detect the different rails, similarly to the non MNNVL case.

In a "per-host" rail configuration, the rails are assumed to connect all the same devices on each host.
To detect if two devices are the same rail, we cannot use the GUID anymore.
The network name and dev index being unreliable accross hosts (they can change), we use the PCIe path to detect that devices are on the same rail when on different hosts.
This is too restrictive when a single device is exposed through different PCIe path, but this approach is conservative: it will not establish connections when it shouldn't.