Add option to pre-post receive work requests on the receiver side
IMPORTANT: The profiling of receive requests is BROKEN in this change and needs
to be fixed!

This change adds an option to pre-post receive work requests (WQEs) on the
receiver side. The option is controlled by the environment variable
`NCCL_IB_PREPOST_RECEIVE_WORK_REQUESTS` which is by default assigned to `0`,
meaning that receive work requests are posted on-demand when a new receive
request is issued during `ncclIbIrecv()` call.

When the option is enabled (by setting the environment variable to `1`),
the receiver first pre-posts receive work requests on all QPs during the
connection establishment phase and additionally, whenever a receive work
request is completed (upon receiving a CQE), another receive work request
is posted on the same QP. This way, there is always a receive work request
available for incoming messages. Note that when a receive work request is
completed, in order to post another receive work request on the same QP, the QP
must be identified first, which is done by looking up the QP using the QP
number in the CQE. The function which looks up for the QP using the QP number
is `ncclIbCommBaseGetQpByQpNum()` and it's optimized not to look for the QP in
all QPs of the communicator, but only in the QPs that belong to the same device
as the device on which the CQE is received.

When configuring the receiver to pre-post receive work requests, it is required
that the matching scheme of receive requests is set to the ID-based scheme
(IB_RECEIVER_SIDE_MATCHING_SCHEME=BY_ID), because when the receiver pre-posts
receive work requests, the work requests are assigned with "dummy" `wr_id`s,
so it's not possible to retrieve the receive request using the `wr_id` in the
CQE.

Because the environment variable is used both in the `connect.cc` file (during
the connection establishment) and also in `p2p.cc` file (in the `ncclIbIrecv()`
and during `ncclIbTest()`), and because is not possible to define a NCCL
parameter in an `.h` file (because of how the files are structures and how the
`NCCL_PARAM` macro is defined) - the environment variable is defined in the
`common.cc` file, and a new member variable `prepostReceiveWorkRequests`
is added to the `ncclIbRecvComm` structure in the `common.h` file, which is
initialized in a new function `ncclIbRecvCommInit()` in the `common.cc` file.
This function is called in the `connect.cc` file after allocating the
`ncclIbRecvComm` structure, in the early stages of the connection
establishment.