Refactor QP selection logic into separate functions
QP selection for send, receive and CTS messages is refactored into separate
functions. This makes it cleaner and easier to understand how QP is selected
and abstracts the implementation of QP selection out of the data path.

The baseComm structure was added with a member that is assigned to the
NCCL_PARAM IbSplitDataOnQps (NCCL_IB_SPLIT_DATA_ON_QPS) because
the NCCL_PARAM cannot be defined in the common.h file because how the
NCCL_PARAM macro is defined.

This macro defines a function (ncclParamIbSplitDataOnQps()) with external
linkage, so including this header in multiple source files causes multiple
definitions of the same function. Therefore the NCCL_PARAM was declared
in connect.cc and assigned in the baseComm as it is used in the function
ncclIbCommBaseGetNqpsPerRequest

The functions added in the common.h are also defined in the common.h
since they are declared as inline to allow compiler optimization.