net: init plugin only once if comms share resources
If a communicator is created with ncclCommSplit and the resources are
shared with the parent communicator (including network connections) then
the network plugin only needs to be initialized once.

If the split communicator defines a traffic class that does not match
the traffic class of the parent we print INFO log allerting users that
the traffic class of the new communicator is ignored and the traffic
class of the parent communicator is used instead. Do the same for
netName

BUG: 5562553