Refactor inconsistency of fifoHead and fifoTail members
Before this fix, the sender and receiver stored the fifoHead and fifoTail
in different structures although the usage should be the same.

In addition, even though sender and receiver use the "CTS FIFO" as a FIFO, the
sender and receiver only care about the "head" of the CTS FIFO and do not
track the tail of it.

The protocol is a CTS-based, so the receiver progresses the CTS FIFO head and
the sender tracks this head. It can be said that the sender follows the tail
of the CTS FIFO but from the sender's perspective, it's the head, and this is
matches also the current naming of the member in the code.

This change puts the fifoHead member in the same place for the sender and
the receiver.

It's possible to also track the fifoTail on the sender and receiver, and
for example check that the number of outstanding requests is not exceeding
the FIFO size but it's not done in the code today and no reason to add this
at this point.