Squash commit for [CI] Use updated gcperf-tools package. Add DB upload
## Summary
We've been using a staging gcperf-tools venv using a version of the package built from a feature branch. The feature we are interested in has now been built and deployed in the main package, so we can switch back to the main venv.
Additionally, adds step to upload perf regression results to an IT-hosted Postgres DB. This provides better historical data storage, and enables us to add our data to gpucomms.nvidia.com

## Implementation
Switch the venv path in eos-config.sh to point to the main venv.
Add gcperf-tools upload step to script.

## Testing
Gitlab CI

## Tracking
https://nvbugspro.nvidia.com/bug/5377685

Switch back to main gcperf-tools venv with updated package

See merge request nccl/nccl!1377