Fix NVTX payloads for commSplit and commShrink
Move the payload initialization to the end of the functions, because the new communicator is only set after the internal group end.

If there the user code creates groups, the communicator will only be created at the corresponding end group.
Thus, we cannot determine the payload information within the comm function in that case.

BUG 5569688