Fix the nElts<->nAllElts confusion
The AllGather and ReduceScatter kernels were in places using nElts where
nAllElts was the correct choice.  When calculating the offsets within
the collective buffer or alignment, nAllElts is generally the value to
use.

These two tend to be the same when everything is perfectly aligned, but
if a collective ends up getting split between channels, they diverge.