Allow ncclIbPostFifo() to be called in different times for retransmission
Reliance on a state (base.fifoHead) not related to the request itself makes
it impossible to call this function when retransmission of a CTS message
is required.

The function ncclIbPostFifo() relies on base.fifoHead for two things:
1. Determining the slot to which the CTS writes to in the CTS FIFO.
2. The "ID" (`idx`) in the CTS message which is a incremental value of
base.fifoHead.

Since both of these values (and many other related to the CTS content itself)
are determined upon the call to ncclIbIrecv() and cannot change through the
life time of the receive request, all of the immutable information is now
changed to be set during the ncclIbIrecv() and when ncclIbPostFifo(), only the
specific fields of the work request that are affected by the specific device
on which the CTS is sent - are updated.

In addition, it's cleaner to rely read and update fifoHead in a single
function (ncclIbIrecv()) instead of splitting it in two functions without
a good reason.