Fix the calculation of end index for work range
getWorkRange was reusing the ncclSymkDevWork structure of the start
index for the calculation of indexHi, whereas it should've been using
the one of the end index.

If there were differences in size between the messages in the group,
this would result in wrong offset calculation, leading to part of the
buffer not being processed (typically the initial part of larger
messages).