Move ncclIbQpInfo and ncclIbConnectionMetadata to connect.h
Resiliency would need to use these structures:

The ncclIbQpInfo structure would be populated by the QP information from the
QPs created for probing.

The ncclIbConnectionMetadata is used as well in the resiliency API. The device
information provided in this structure is being stored in the resiliency
context so each device would "know" to which device it's connected to on the
remote side.