enqueue: fix a2a batching
The previous fix did not effectively ensure that we would separate
p2p operations by epochs of 8 p2p rounds: if we skipped round 8 for
example, because we had nothing to send/recv on that round, we would
not create a new batch and we would continue to add to the current
wip batch.

This fix creates a new batch if our p2p round is not in the same epoch
as the operations currently in the batch.

BUG 5589277