init/a2a: do not split single node NVL into groups
The purpose of groups in the a2a scheduling is to guarantee the PXN aggregation factor.
Groups of size of max 8 are required to guarantee that GPUs on the same host will schedule remote peers on the same channels.

In the case of single node, no PXN is required.
In that case, we choose a "close-ranks" first approach (groupSize = maxLocalRanks).

BUG 5589277