a2a: do not restrict the channels per net peer for pure NVL
In the case of pure NVLD, the latency of the fabric is so slow and the
message rate so high that having many channels per net peer, even in the
case of the a2a doesn't lead to any performance impact for the a2a.

Therefore, keeping a very high value will guarantee good performance in
the case of sparser send/recv patterns.

BUG 5593828