Allow ncclIbMultiSend to be called in different times for retransmission
The problem was that the sendReqs pointers array was reset as soon as
ncclIbMultiSend was complete and upon replay, the sendReqs array was not
valid.

Before this change, the sendReqs were nullified (using memset(0)) right after
the call to ncclIbMultiSend(). This memset is not only for "debug" as the
comment above states but also it has functional purposes. Without this memset,
following send requests made by user are not accepted because of this line in
the ncclIbIsend() function:

if (reqs[r] != NULL || slots[r].tag != tag) continue;

Having this memset() call is ok/needed for two reasons:
1. After the call to ncclIbMultiSend, the array of pointers is not required
anymore (as it is assumed that there are no failures and no need for
retransmit anything).
2. In order for the ncclIbIrecv() to accept new send requests as soon as
the previous (multi-)send request was posted (not even completed!)

Note that the line that resets the "slot" buffer:

memset((void*)slots, 0, sizeof(struct ncclIbSendFifo))

is, indeed, not really needed for functionality, and maybe was added only for
debug as mentioned in the comment above but it hides a potential race bug.
Consider the following case:
1. The called for ncclIbMultiSend() completed
2. The thread was deallocated from the CPU.
3. The local NIC completed the (multi-)send request.
4. Receiver completed the (multi-)recv request.
5. Receiver wrote a new CTS to the same slot[1]
6. The local thread returns and overwrites/resets the slot with zeros.

This change removes both "reset" calls from ncclIbIsend(). The reason for that
is that when there is a port failure, and a (multi-)send request needs to be
retransmitted, the formatting of the work requests is relying on the
availability of the information in the slots buffer and the validity of the
sendReqs array.

In order to maintain the functionality of the sendReqs reset, the array of
pointers is reset when the (multi-)send request completes.

**IMPORTATNT**
Postponing the reset of the (multi-)send request to the time when it completes
can theoretically delay the time in which the call for ncclIbIsend() can
succeed. Fixing this can be done by "doubling" the sendReqs pointers array and
use kind of "generation" solution, each slot alternates between using the
sendReqs from generation 0 and from generation 1.