Remove events tracking for a CTS message
Before this change, when a Clear-To-Send (CTS) message was posted, events were
used to track the completion of the message (when it was configured to
generate a completion, i.e., signaled) by appending an event to the
`ncclIbRequest` and decrementing an event upon completion of the CTS.

There is no real need for events tracking for a CTS and this change removes
this tracking entirely.

Additionally, removing events tracking for a CTS allows easier replay of a CTS
when port failure requires a CTS to be replayed. When a CTS was tracked with
events, the `initialEventCounters` value was incremented during replay which
makes is cumbersome to decrement it before replay.