Squash commit for Add functions to access device pointers
## Summary
This adds 2 new host functions to the API related the symmetric memory registration.
The 2 functions allow access to the Multimem and LSAPeer pointers.
This was previously accessible on the device side via the NCCL device API, and this MR enables access on the host side as well.

## Implementation

The ncclWindow_t type contains the information necessary to access these pointers.
But ncclWindow_t is a device pointer handle inaccessible from the host.
First step is to use the NCCL shadow pool to look up the corresponding host side window.
Then we use the existing functions `ncclDevrGetLsaTeamPtrMC` i.e. `ncclDevrGetLsaRankPtr` to obtain the pointers.
They are still device pointer, but now they are accessible from the host to be passed into a kernel.

The new functions are not considered thread safe, and are also not designed to be SOL.
But they are useful. The returned pointer have the lifetime of the associated window and communicator, whatever is shorter.

## Testing

This MR introduces new tests for CI.
They compare the pointers from the host side API with the pointers obtained in device code.
Manually tested on Pre-Tyche GB200 with NVLINK available.

## Tracking
https://nvbugspro.nvidia.com/bug/5466255
https://jirasw.nvidia.com/browse/GP-417

Adding Host functions to get peer pointer or multimem pointer.

See merge request nccl/nccl!1317