Squash commit for Disable use of Linux Abstract Sockets for IPC over UDS
## Summary
Disable the use of the Linux Abstract Socket mechanism for IPC UDS Sockets.
A customer reported that this feature prevented containers running on the same server from communicating with each other.

## Implementation
Previously the code used a hard-coded define `USE_ABSTRACT_SOCKET` and code to enable the Linux Abstract Socket feature.
This change now makes that controllable via a new environment variable:

   `NCCL_IPC_USE_ABSTRACT_SOCKET`

which is defaulted to 0.

There is a small risk that stale UDS files will be left in `/tmp` if the application crashes.
But the code does an `unlink()` before creating a new UDS socket to counter this.

## Testing
Testing was done on a B300 with 8 GPUs testing both enabled and disabled paths.

## Tracking
BUG 5529136

Disable use of Linux Abstract Sockets for UDS

This was preventing communucation between containers running on
the same server.

Now controlled by a new env var:

   NCCL_IPC_USE_ABSTRACT_SOCKET

which defaults to 0

BUG 5529136

See merge request nccl/nccl!1455