Squash commit for Allow ncclDebugLog to print messages longer than 1024 bytes
## Summary
This change removes a limitation on log line length in NCCL.  This can be very useful for debugging and has been requested by users:
- https://github.com/NVIDIA/nccl/pull/825
- https://github.com/NVIDIA/nccl/pull/1215

## Implementation
Currently the NCCL logging code formats the line to be printed in a temporary buffer, and then logs it to the output stream by calling `fwrite`.  This change replaces the two-step process.  Instead of copying the entire formatted string into the temporary buffer, it used the temporary buffer to create a prefixed format string by prepending NCCL's prefix to the user-supplied format.  This format string is then supplied to `vfprintf`.  This means that values are printed directly to the output, without an intermediate buffer. 

## Testing
This is a patch that we used in our NCCL deployments at Enfabrica.

## Tracking
Patch adapted from https://github.com/NVIDIA/nccl/pull/1215

Allow ncclDebugLog to print messages longer than 1024 bytes

See merge request nccl/nccl!1424