Squash commit for [CI] Migrate IPP6 jobs to BCM11 cluster
## Summary
IPP6 cluster is undergoing an upgrade to BCM 11. To upgrade we need to setup a new cluster adjacent to the old one and begin moving nodes over. We've setup enough nodes in the new cluster to facilitate NCCL CI. Once CI is moved over with this MR, we can move the rest of the CI nodes over to restore the capacity.

After CI is moved over, we will just need to coordinate with the devs to fully deprecate the old cluster and get all nodes migrated into BCM 11.

## Implementation
Updates IPP6 jobs to use new Gitlab runner tag that lands on BCM 11 cluster

## Testing
Gitlab CI

## Tracking
https://nvbugspro.nvidia.com/bug/5539702

Change IPP6 to BCM11 cluster

See merge request nccl/nccl!1448