Squash commit for Fix VERSION logging garbage
## Summary
A recent change (MR !1424) has caused VERSION level logs to log garbage instead. This fixes it.

## Implementation
The previous change was implemented by including the `fmt` into `buffer`, without applying the varargs yet. Then instead of writing the completed buffer to the output stream, `buffer` was passed to `fprintf` with the varargs. This way, the size of the arguments would not count against the statically sized `buffer`. However, the place where `fmt` was included in `buffer` does not have a branch for VERSION. This fixes adds an else which just copies `fmt` to `buffer`.

## Testing
Ran it before the fix. No VERSION log line was observed. Ran it after the fix and the VERSION log line appeared as expected.

## Tracking
None.

Fix VERSION logs logging garbage

See merge request nccl/nccl!1503