TMA: add support for symmetric kernels
TMA stands for Tensor Memory Accelerator. This component is an async
data movement engine that equips GPUs since Hopper generation. TMA can
do bulk loads (from local/remote Global Memory to Shared Memory) and
stores (from Shared Memory to local/remote Global Memory) asynchronously
w.r.t. the issuing thread.

TMA operations are warp uniform, meaning all threads in the warp invoke
TMA instructions with the same arguments producing a single TMA
instruction as result or a single thread in the warp is elected to issue
TMA operations on behalf of all threads.

TMA operations completion is always local. For loads, completion is
checked using a dedicate mbarrier object allocated in shared memory and
updated by the TMA engine with the amount of data transferred from the
original request. When the barrier is released, data is in shared memory
and is ready to be consumed by threads in the warp.

For stores, completion is checked using groups. A group of TMA stores
can be commited (flushed) and waited upon. When the group wait completes
the shared memory buffers can be reused.

The following symmetric kernels are supported w/ TMA:

1. AllGather_ST and AllGather_STMC
2. ReduceScatter_LD
3. AllReduce_RSxLD_AGxST

BUG: 5312943