Add port-failure emulation (PFE) library and CLI tool
The PFE library allows to emulate port failure on devices. It's done by
inserting steering rules into the device that drop all traffic to/from the
port. The CLI tool allows easy usage of the library and provides a simple
interface to trigger the emulation.

The library and the CLI tool currently work only on devices in RoCE/SPCX mode

The build of PFE library is done conditionally only if the dependencies are met.
For example libibverbs and libmlx5 are found.

The PFE library respects log levels using PFE_DEBUG env variable.

The build of PFE library is done using CMake or make. The build is done only
if the dependencies are met. For example libibverbs and libmlx5 are found.

Less important notes:
* pfe_internal.h includes <ctype.h> for `toupper()` function call