Squash commit for [Perf Regression] Make segments arg more granular based on nodes
## Summary
Perf regression testing on pre-tyche has been unstable, occasionally showing 30% regressions. I discovered this is because --segment needs to be specified to ensure the requested nodes are all in the same nvlink domain. This argument is being used for weekly tests, but it is also needed in our pre-merge, post-merge, and nightly testing across 4 nodes.

## Implementation
Adds logic to add --segment for all pre-tyche jobs, with the value differing based on the number of nodes (which is determined by the pipeline type)

## Testing
Gitlab CI

## Tracking
https://jirasw.nvidia.com/browse/SGCOMDEV-221

Fix segments

See merge request nccl/nccl!1513