Squash commit for proxy: add acquire/release logic to free ops
## Summary
This MR adds a missing acquire/release logic to the proxy op.

## Implementation
When the proxy thread is done with ops, it releases them to the main thread to be recycled.
This part of the code was missing an acquire/release logic that is added in this MR.
This ensures the needed store ordering in memory for the next pointers values to be visible by the main thread

## Testing
done with a reduced `MAX_OPS_PER_PEER` to 128 instead of 1024 to expose the code path more often.
- :heavy_check_mark: `gcp-iad` 256 GPUS with NCCL_P2P_DISABLE=1 and NCCL_SHM_DISABLE=1
- :heavy_check_mark: `gcp-iad` 256 GPUS with NCCL_MNNVL_DISABLE=1 (to cover PXN)
- performance sweep with and without the patch, see below

## Tracking

- https://nvbugs/5562698 [GGL][Stellaris][GB200] 4-domain NCCL all-to-all often hangs


![time_comparison](/uploads/336858dfc7db5594235562ba0d8688c1/time_comparison.png)
![busbw_comparison](/uploads/79a63f736a58b116e3227085453a8df9/busbw_comparison.png)

proxy: add acquire/release logic to free ops

When the proxy thread is done processing ops, it returns them to the
main thread.
The new acquire/release logic guarantees that the operations done before
such as linking the ops together (with next) will be visible on the main
thread.

BUG 5562698

See merge request nccl/nccl!1517