Squash commit for [CI] Make perf regression use feature branch as comparison
## Summary
All perf-regression jobs compare results to the latest result on master. On feature branches this can mean that a regression anywhere in the branch can block all future MRs to that branch. Instead the MRs targeting that branch should compare themselves to the latest result on that branch only.

## Implementation
Set the golden branch to the target branch of the MR if available. If not, which means the pipeline is post-merge or nightly, use the current branch as the comparator.

## Testing
Gitlab CI

## Tracking
N/A

Make golden branch dynamic on target branch

See merge request nccl/nccl!1551