Squash commit for support deviceAPI versioning
## Summary
This MR introduces versioning for the Dev API so that future updates can support backwards-compatibility. 

This is a **breaking change**: applications compiled with 2.28 are not compatible with 2.29. This one-time breaking change allows for non-breaking changes in the future.

## Implementation
User-facing structs (``ncclDeviceProperties_t`` and ``ncclDevComm_t``) have three key fields added:
* version
* magic
* size

These fields should be initialized with a custom user-facing ``INITIALIZER`` macro. ``magic`` is used to verify the fields are properly initialized. In the future, we can user the version/size to selectively access fields based on the compile-time version.

## Testing
Added ``devCommCreate`` apitests to check for success when properly initialized and failure otherwise.

## Tracking

https://nvbugspro.nvidia.com/bug/5490042

deviceAPI: add version to commProperties and devCommRequirements

See merge request nccl/nccl!1464