Refactor and optimize the size assignment of send requests in ncclIbMultiSend
In terms of functionality, this change does not alter the behavior of the code.
Even though that with the new code, when `nreqs==1`, the buffer into which the
size of the single send request was assigned, not used - it does not affect the
correctness of the code.

The previous implementation used an `if (nreqs > 1)` check which introduced
unnecessary branch overhead and risked branch misprediction, especially when
`nreqs` was 1.

This change uses a single `for` loop. , eliminating the conditional branch and
simplifying the code for the compiler, leading to better performance
predictability and potential vectorization.

In addition, this change will help future refactoring of the code.

Note how many times there is `for (int r=0; r<nreqs ; r++)` loop in this file.