Refactor ncclIbMultiSend to calculate chunkSizes only once
In the function ncclIbMultiSend, the chunkSizes array is calculated multiple
times. When selective retransmission will added, it will be necessary to have
another calculate of the chunk sizes so this change aims to refactor the code
to calculate chunkSizes only once per request.

In addition, the ncclIbRequest::offset member is not really required and was
used only as a helper to update the request during posting but that can be a
local variable inside ncclIbMultiSend().

No need to assign a null request that causes a seg-fault in line 256 after the change