Fix graph search failure on A100 systems with 1 NIC per GPU.
A100 graph search relies only on balanced tree, which may need to
come back through a SYS or PHB path on platforms with 1 NIC per GPU.

The restriction of path types (to accelerate search) prevents that
path to be found.

Disable the restriction on A100 platforms.

Also add the Cineca platform to the unit tests.

BUG 5635951
Github issue #1876