Squash commit for Use ARCHIVER_STAGING_DIR to avoid using local /tmp on runner nodes
## Summary
Gitlab executor processes have been using /tmp on runner nodes as a staging directory for unzipping artifacts. On some clusters, Pre-Nyx in particular, CI jobs have been failing because the disk space on the runner nodes were getting full and the executors could no longer write to `/tmp`. To prevent this, we are now moving the staging directory to LustreFS space on the clusters instead. 

## Implementation
Set `ARCHIVER_STAGING_DIR` on every job to pass a directory on LustreFS to be used for staging.
https://docs.gitlab.com/ci/runners/configure_runners/#staging-directory

## Testing
Passing pipeline should tell us that no functionality was broken. Manually inspected the LustreFS directory to confirm it's being used as the staging space.

added ARCHIVER_STAGING_DIR to prenyx jobs

See merge request nccl/nccl!1614