Squash commit for [CI] Add comparison for release branches to master
## Summary
Perf regression currently functions in the following was
- When a user opens an MR, the performance on their branch is compared to the MR's target branch
- When the MR is merged, a post-submit pipeline runs to compare the branch's performance to the last nightly result
- On a nightly/weekly basis, the branch re-runs performance tests, checks for regressions, then saves the results as new baseline for the branch

The gap in this lies on release branches such as v2.29. There is no automatic comparison between a release branch like 2.29 and master. This can only be found if a user manually opens an MR for the release branch and runs perf-regression testing.  Instead what should happen is that the scheduled pipelines for the release branches should automatically add an additional perf-regression comparison against the laster results from master. This will impact the error code of the job, as well as file an additional NvBug should a regression be detected.

## Implementation
- Add variable COMPARE_TO_BASELINE that is automatically set if the pipeline is scheduled and matches the pattern of a release branch (can also be passed in manually)
- If COMPARE_TO_BASELINE is set, the job will run an additional regression check using master's results as the base values.
- Will file up to two NvBugs if there are regressions on both the release branch (compared to the last run) and master
- Refactored script to use functions to implement the above with minimal additional code

## Testing
Gitlab CI

## Tracking
https://jirasw.nvidia.com/browse/SGCOMDEV-225

Add comparison for release branches to master

See merge request nccl/nccl!1624