Fix return codes and add WARNS for invalid usage cases.
Return proper error codes when we cannot register the window, and
return a NULL window.

Use CommCheck and PtrCheck functions instead of custom checks.

Also, in general, the top layer is reponsible for checking arguments.
Further checks should either be removed or return an internal error
(and not take a lot of code space).