Store the ECE set on a QP within the QP itself for restore purposes
When a QP is set with ECE (whether ECE is supported or not), store the ECE into
the QP itself.

When a port failure happens and later on recovers, the QP should be restored.
In order to avoid full connection establishment protocol with the remote side,
the ECE is stored so each side can set the ECE without further handshake.