Refactor ncclIbRtrQp to use attributes structure as input
Notes:

1. this changes solves the problem where the MTU is being reduced and
assigned to the remote device:

remDevInfo->mtu = std::min(remDevInfo->mtu, ibDev->portAttr.active_mtu);

This assignment to the remDevInfo, was done because the ncclIbRtrQp accepted
only the remDevInfo structure so this was the way to pass the reduced MTU.

2. The FifoTC feature (NCCL_IB_FIFO_TC) is respected for "data-transferring
QPs" only.

When propagating this change in ncclIbRtrQp to the resiliency module, resiliency
module will not use/respect the FifoTC feature, because the parameter is not
accessible in the resiliency module outside of connect.cc.