Add ncclIbInitQp() function
The change separates the transition of a QP to Init state from within the
function that creates it (ncclIbCreateQp()). This is done in order that
when a QP fails, it would be possible to reuse the function ncclIbInitQp()
to initialize the QP.

The function gets only the QP (ncclIbQp) and assumes that the QP was
initialized with the init attributes required (ncclIbQp::initAttr).