Align QP configuration functions to the same structure and naming
Except for ncclIbQpCreate() function, all functions have the same input
arguments.

The reason that ncclIbQpCreate() function does not follow is because the
QP create attributes are not stored in the QP. This is because QP is never
destroyed during restoration upon port recovery.

After this change, all functions follow the ncclIb<object><action> pattern. In
this case ncclIbQp<action>.