Squash commit for Fix description of NVLS_ENABLE
## Summary

Fix the description of the NVLS_ENABLE parameter to match current implementation.

## Implementation

Investigated the source. Kamil has tested to confirm that the previous description was wrong.

## Testing

Discussed with Addy and Kamil.

Built on gc00 and verified that it builds without errors.

## Tracking

[\[B\] 5741374: \[NCCL 2.29\] Test is not failed as expected with NCCL_NVLS_ENABLE=1 on NVLS not supported platforms.](https://nvbugspro.nvidia.com/bug/5741374)

Fix description of NVLS_ENABLE

Behavior of NVLS_ENABLE has changed recently and the description needs
updating in regards to how it handles failures.

See merge request nccl/nccl!1733